+++ zoom/models.py
@@ -120,6 +120,11 @@ class Members(RecordStore):
         )
 
 
+def get_members(db=None):
+    db = db or zoom.get_db()
+    return Members(db)
+
+
 class Subgroup(Record):
     pass
 
@@ -242,6 +247,17 @@ class Group(Record):
         membership = members.first(group_id=self._id, user_id=user._id)
         if not membership:
             members.put(Member(group_id=self._id, user_id=user._id))
+            audit('add member', self.name, user.username)
+
+    def remove_user(self, user):
+        """Remove a user from a group"""
+        store = self.get('__store')
+        members = Members(store.db)
+        membership = members.first(group_id=self._id, user_id=user._id)
+        if membership:
+            cmd = 'delete from members where group_id=%s and user_id=%s'
+            store.db(cmd, self._id, user._id)
+            audit('remove member', self.name, user.username)
 
     @property
     def roles(self):